Split the IntermediateFile file for better OO design
- getIntermediateFile went into its own factory file
- the subclasses of IntermediateFile get their own file
- IntermediateFile.js is now just a simple abstract class
